Prepare Render deploy and improve responsive UI

+++ build.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
 
-pip install -r requirements.txt
+set -o errexit
+
+if [ -f requirements-render.txt ]; then
+  pip install -r requirements-render.txt
+else
+  pip install -r requirements.txt
+fi
+
 python manage.py collectstatic --noinput
 python manage.py migrate
